# Use class cache in factory