This avoid a lot of ``class_exists`` calls, saving some precious milliseconds !