@EmmanuelVella 2015-01-26
Sure ! I made a quick benchmark on my local machine, on a symfony2 prod environment :

 | Before | After
---- | ------------ | -------------
Calls | 5332 | 81
Total time | 485 ms | 460 ms