@lsmith77 2015-01-26
while you are at it .. I always wondered if https://github.com/jackalope/jackalope/pull/272/files#diff-710f3ac2ecf6432c840eacf5dba8209eR60 is causing a noticeable amount of overhead. ie.if it would make sense to hardcode a few cases (like 1, 2, 3, 4 constructor parameters).