@EmmanuelVella 2015-01-27
Here is a benchmark with 5 000 (realist) and 10 000 000 iterations : 

      | 5 000 | 10 000 000
---- | ------------ | -------------
Without reflection cache | ~14 ms | 25672 ms
With reflection cache | ~6 ms | 20405 ms