@EmmanuelVella 2015-01-27
I just did another benchmark for the class cache I added in this PR : 

      | 5 000 | 10 000 000
---- | ------------ | -------------
Without class cache | ~24 ms | 52312 ms
With class cache | ~6 ms | 20405 ms