@lsmith77 2015-02-04
@dbu fine for me to merge this .. 

@EmmanuelVella did you also try to benchmark if hardcoding different permutations of constructor parameter counts performs faster than using Reflection for all cases?