@lsmith77 2015-02-04
ie something like this (I did a quick check and 6 seems to be the maximum number of constructor params the use):
```
diff --git a/src/Jackalope/Factory.php b/src/Jackalope/Factory.php
index d9b0a0c..89ed384 100644
--- a/src/Jackalope/Factory.php
+++ b/src/Jackalope/Factory.php
@@ -28,12 +28,25 @@ class Factory implements FactoryInterface
             array_unshift($params, $this);
         }
 
-        if (! count($params)) {
-            return new $name;
-        }
-
-        $class = new ReflectionClass($name);
+        switch (count($params)) {
+            case 0:
+                return new $name();
+            case 1:
+                return new $name($params[0]);
+            case 2:
+                return new $name($params[0], $params[1]);
+            case 3:
+                return new $name($params[0], $params[1], $params[2]);
+            case 4:
+                return new $name($params[0], $params[1], $params[2], $params[3]);
+            case 5:
+                return new $name($params[0], $params[1], $params[2], $params[3], $params[4]);
+            case 6:
+                return new $name($params[0], $params[1], $params[2], $params[3], $params[4], $params[5]);
+            default:
+                $class = new ReflectionClass($name);
 
-        return $class->newInstanceArgs($params);
+                return $class->newInstanceArgs($params);
+        }
     }
 }
```