@EmmanuelVella 2015-02-05
Here are the results : 

      | 5 000 | 10 000 000
---- | ------------ | -------------
With ReflectionClass  | ~12 ms | 23588 ms
With switch| ~10 ms | 16981 ms

IMHO the ``switch`` makes the code more complicated and do not provide a noticeable performance improvement (in a normal use case).