@dbu 2015-02-06
thanks a lot! 
we started thinking about jackalope 2.0 - i think the factory is one place to think again. i added some ideas to https://github.com/jackalope/jackalope/wiki/2.0#factory